ml/requirements.txt
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
matplotlib>=3.8
tensorflow-cpu>=2.16

